+++ examples/06_custom_tech/user_defined_model/paper_mill.py
@@ -3,8 +3,8 @@ import openmdao.api as om
 import numpy_financial as npf
 from attrs import field, define
 
-from h2integrate.core.utilities import BaseConfig, CostModelBaseConfig, merge_shared_inputs
-from h2integrate.core.model_baseclasses import CostModelBaseClass
+from h2integrate.core.utilities import BaseConfig, merge_shared_inputs
+from h2integrate.core.model_baseclasses import CostModelBaseClass, CostModelBaseConfig
 
 
 n_timesteps = 8760