+++ examples/25_pyomo_optimized_dispatch/run_pyomo_optimized_dispatch.py
@@ -5,7 +5,7 @@ from h2integrate.core.h2integrate_model import H2IntegrateModel
 
 
 # Create an H2Integrate model
-model = H2IntegrateModel("pyomo_heuristic_dispatch.yaml")
+model = H2IntegrateModel("pyomo_optimized_dispatch.yaml")
 
 demand_profile = np.ones(8760) * 100.0