+++ h2integrate/converters/hydrogen/geologic/h2_well_subsurface_baseclass.py
@@ -1,9 +1,9 @@
 import openmdao.api as om
 from attrs import field, define
 
-from h2integrate.core.utilities import BaseConfig, CostModelBaseConfig
+from h2integrate.core.utilities import BaseConfig
 from h2integrate.core.validators import contains
-from h2integrate.core.model_baseclasses import CostModelBaseClass
+from h2integrate.core.model_baseclasses import CostModelBaseClass, CostModelBaseConfig
 from h2integrate.tools.inflation.inflate import inflate_cepci