+++ h2integrate/storage/hydrogen/tank_baseclass.py
@@ -1,8 +1,7 @@
 import openmdao.api as om
 from attrs import field, define
 
-from h2integrate.core.utilities import CostModelBaseConfig
-from h2integrate.core.model_baseclasses import CostModelBaseClass
+from h2integrate.core.model_baseclasses import CostModelBaseClass, CostModelBaseConfig
 
 
 class HydrogenTankPerformanceModel(om.ExplicitComponent):